refactor: move graph results
graph result represents an application specific result.
