+++ src/ecalc_cli/commands/run.py
@@ -5,8 +5,8 @@ import libecalc.common.time_utils
 import libecalc.version
 import typer
 from libecalc.application.energy_calculator import EnergyCalculator
+from libecalc.application.graph_result import GraphResult
 from libecalc.common.run_info import RunInfo
-from libecalc.core.graph_result import GraphResult
 from libecalc.infrastructure.file_utils import OutputFormat, get_result_output
 from libecalc.presentation.yaml.model import YamlModel
