+++ src/libecalc/application/graph_result.py
