+++ src/tests/libecalc/dto/utils/test_aggregators.py
@@ -4,12 +4,12 @@ from typing import List
 import pandas as pd
 from libecalc import dto
 from libecalc.application.energy_calculator import EnergyCalculator
+from libecalc.application.graph_result import GraphResult
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import (
     TimeSeriesFloat,
     TimeSeriesStreamDayRate,
 )
-from libecalc.core.graph_result import GraphResult
 from libecalc.core.result.emission import EmissionResult
 from libecalc.dto.result.emission import PartialEmissionResult
 from libecalc.dto.utils.aggregators import aggregate_emissions
