+++ src/tests/libecalc/integration/test_all_energy_usage_models.py
@@ -1,6 +1,6 @@
 import pytest
 from libecalc.application.energy_calculator import EnergyCalculator
-from libecalc.core.graph_result import EnergyCalculatorResult, GraphResult
+from libecalc.application.graph_result import EnergyCalculatorResult, GraphResult
 
 
 @pytest.fixture
